+++ car _rental/model/Rent.cs
@@ -2,6 +2,6 @@ namespace katas.car_rental
 {
     public interface Rent
     {
-         int Amount();
+         double Amount();
     }
 }
